+++ src/main/java/org/folio/rest/impl/DataImportImpl.java
@@ -8,24 +8,22 @@ import io.vertx.core.Vertx;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
 import org.apache.http.HttpStatus;
-import org.folio.rest.jaxrs.model.DataImportUploadFilePostMultipartFormData;
 import org.folio.rest.jaxrs.model.DefinitionCollection;
-import org.folio.rest.jaxrs.model.File;
+import org.folio.rest.jaxrs.model.FileDefinition;
 import org.folio.rest.jaxrs.model.UploadDefinition;
 import org.folio.rest.jaxrs.resource.DataImport;
 import org.folio.rest.tools.utils.TenantTool;
-import org.folio.service.FileService;
-import org.folio.service.FileServiceImpl;
-import org.folio.service.UploadDefinitionService;
-import org.folio.service.UploadDefinitionServiceImpl;
+import org.folio.service.file.FileService;
+import org.folio.service.file.FileServiceImpl;
+import org.folio.service.upload.UploadDefinitionService;
+import org.folio.service.upload.UploadDefinitionServiceImpl;
 import org.folio.util.DataImportHelper;
 
 import javax.ws.rs.NotFoundException;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import java.util.Date;
+import java.io.InputStream;
 import java.util.Map;
-import java.util.Optional;
 
 public class DataImportImpl implements DataImport {
 
@@ -36,7 +34,7 @@ public class DataImportImpl implements DataImport {
   public DataImportImpl(Vertx vertx, String tenantId) {
     String calculatedTenantId = TenantTool.calculateTenantId(tenantId);
     this.uploadDefinitionService = new UploadDefinitionServiceImpl(vertx, calculatedTenantId);
-    this.fileService = new FileServiceImpl(vertx, calculatedTenantId);
+    this.fileService = new FileServiceImpl(vertx, calculatedTenantId, this.uploadDefinitionService);
   }
 
   @Override
@@ -61,7 +59,7 @@ public class DataImportImpl implements DataImport {
                                             Map<String, String> okapiHeaders,
                                             Handler<AsyncResult<Response>> asyncResultHandler,
                                             Context vertxContext) {
-    vertxContext.runOnContext(v -> {
+    vertxContext.runOnContext(c -> {
       try {
         uploadDefinitionService.getUploadDefinitions(query, offset, limit)
           .map(definitions -> new DefinitionCollection()
@@ -83,14 +81,12 @@ public class DataImportImpl implements DataImport {
                                                           Map<String, String> okapiHeaders,
                                                           Handler<AsyncResult<Response>> asyncResultHandler,
                                                           Context vertxContext) {
-    vertxContext.runOnContext(v -> {
+    vertxContext.runOnContext(c -> {
       try {
         entity.setId(definitionId);
         uploadDefinitionService.updateUploadDefinition(entity)
-          .map(updated -> updated ?
-            PutDataImportUploadDefinitionByDefinitionIdResponse.respond200WithApplicationJson(entity) :
-            buildUploadDefinitionNotFound(definitionId)
-          )
+          .map(PutDataImportUploadDefinitionByDefinitionIdResponse::respond200WithApplicationJson)
+          .map(Response.class::cast)
           .otherwise(DataImportHelper::mapExceptionToResponse)
           .setHandler(asyncResultHandler);
       } catch (Exception e) {
@@ -104,7 +100,7 @@ public class DataImportImpl implements DataImport {
   public void getDataImportUploadDefinitionByDefinitionId(String definitionId, Map<String, String> okapiHeaders,
                                                           Handler<AsyncResult<Response>> asyncResultHandler,
                                                           Context vertxContext) {
-    vertxContext.runOnContext(v -> {
+    vertxContext.runOnContext(c -> {
       try {
         uploadDefinitionService.getUploadDefinitionById(definitionId)
           .map(optionalDefinition -> optionalDefinition.orElseThrow(() ->
@@ -121,18 +117,14 @@ public class DataImportImpl implements DataImport {
   }
 
   @Override
-  public void postDataImportUploadFile(String jobExecutionId, String uploadDefinitionId,
-                                       DataImportUploadFilePostMultipartFormData entity, Map<String, String> okapiHeaders,
-                                       Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
+  public void postDataImportUploadDefinitionFile(FileDefinition entity, Map<String, String> okapiHeaders,
+                                                 Handler<AsyncResult<Response>> asyncResultHandler,
+                                                 Context vertxContext) {
     vertxContext.runOnContext(c -> {
       try {
-        File file = new File()
-          .withCreateDate(new Date())
-          .withJobExecutionId(jobExecutionId)
-          .withUploadDefinitionId(uploadDefinitionId)
-          .withLoaded(false);
-        fileService.addFile(file, entity)
-          .map(buildFileCreateResponse(uploadDefinitionId))
+        uploadDefinitionService.addFileDefinitionToUpload(entity)
+          .map(PostDataImportUploadDefinitionFileResponse::respond201WithApplicationJson)
+          .map(Response.class::cast)
           .otherwise(DataImportHelper::mapExceptionToResponse)
           .setHandler(asyncResultHandler);
       } catch (Exception e) {
@@ -142,23 +134,44 @@ public class DataImportImpl implements DataImport {
     });
   }
 
-
   @Override
-  public void deleteDataImportUploadFileByFileId(String fileId, Map<String, String> okapiHeaders,
-                                                 Handler<AsyncResult<Response>> asyncResultHandler,
-                                                 Context vertxContext) {
+  public void deleteDataImportUploadDefinitionFileByFileId(String fileId, String uploadDefinitionId,
+                                                           Map<String, String> okapiHeaders,
+                                                           Handler<AsyncResult<Response>> asyncResultHandler,
+                                                           Context vertxContext) {
     try {
-      vertxContext.runOnContext(c -> {
-        asyncResultHandler.handle(Future.succeededFuture(
-          DeleteDataImportUploadFileByFileIdResponse
-            .respond204WithTextPlain("File was successfully deleted")));
-      });
+      vertxContext.runOnContext(c -> fileService.deleteFile(fileId, uploadDefinitionId)
+        .map(deleted -> deleted ?
+          DeleteDataImportUploadDefinitionFileByFileIdResponse.respond204WithTextPlain(
+            String.format("File with id: %s deleted", fileId)) :
+          buildUploadDefinitionNotFound(uploadDefinitionId)
+        )
+        .otherwise(DataImportHelper::mapExceptionToResponse)
+        .setHandler(asyncResultHandler));
     } catch (Exception e) {
       LOG.error("Error during file delete", e);
       asyncResultHandler.handle(Future.succeededFuture(
-        DeleteDataImportUploadFileByFileIdResponse.
+        DeleteDataImportUploadDefinitionFileByFileIdResponse.
           respond500WithTextPlain(Response.Status.INTERNAL_SERVER_ERROR.getReasonPhrase())));
     }
+  }
+
+  @Override
+  public void postDataImportUploadFile(String uploadDefinitionId, String fileId, InputStream entity,
+                                       Map<String, String> okapiHeaders,
+                                       Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
+    vertxContext.runOnContext(c -> {
+      try {
+        fileService.uploadFile(fileId, uploadDefinitionId, entity, okapiHeaders)
+          .map(PostDataImportUploadFileResponse::respond200WithApplicationJson)
+          .map(Response.class::cast)
+          .otherwise(DataImportHelper::mapExceptionToResponse)
+          .setHandler(asyncResultHandler);
+      } catch (Exception e) {
+        asyncResultHandler.handle(Future.succeededFuture(
+          DataImportHelper.mapExceptionToResponse(e)));
+      }
+    });
   }
 
   private Response buildUploadDefinitionNotFound(String definitionId) {
@@ -167,12 +180,5 @@ public class DataImportImpl implements DataImport {
       .type(MediaType.TEXT_PLAIN)
       .entity(String.format("Upload Definition with id '%s' not found", definitionId))
       .build();
-  }
-
-  private Response buildFileCreateResponse(String uploadDefinitionId) {
-    Optional<UploadDefinition> def = uploadDefinitionService.getUploadDefinitionById(uploadDefinitionId).result();
-    return def.map(PostDataImportUploadFileResponse::respond201WithApplicationJson)
-      .orElseGet(() -> PostDataImportUploadFileResponse
-        .respond404WithTextPlain("Upload Definition with id " + uploadDefinitionId + " not found"));
   }
 }