+++ src/main/java/org/folio/service/upload/UploadDefinitionService.java
@@ -1,6 +1,7 @@
-package org.folio.service;
+package org.folio.service.upload;
 
 import io.vertx.core.Future;
+import org.folio.rest.jaxrs.model.FileDefinition;
 import org.folio.rest.jaxrs.model.UploadDefinition;
 
 import java.util.List;
@@ -36,7 +37,7 @@ public interface UploadDefinitionService {
    * @param uploadDefinition UploadDefinition to save
    * @return future with generated id
    */
-  Future<String> addUploadDefinition(UploadDefinition uploadDefinition);
+  Future<UploadDefinition> addUploadDefinition(UploadDefinition uploadDefinition);
 
   /**
    * Updates UploadDefinition with given id
@@ -44,7 +45,7 @@ public interface UploadDefinitionService {
    * @param uploadDefinition UploadDefinition to update
    * @return future with true is succeeded
    */
-  Future<Boolean> updateUploadDefinition(UploadDefinition uploadDefinition);
+  Future<UploadDefinition> updateUploadDefinition(UploadDefinition uploadDefinition);
 
   /**
    * Deletes UploadDefinition by id
@@ -53,5 +54,13 @@ public interface UploadDefinitionService {
    * @return future with true is succeeded
    */
   Future<Boolean> deleteUploadDefinition(String id);
+
+  /**
+   * Add File Definition into Upload Definition
+   *
+   * @param fileDefinition - new file definition to add
+   * @return future with updated UploadDefinition
+   */
+  Future<UploadDefinition> addFileDefinitionToUpload(FileDefinition fileDefinition);
 
 }