+++ NEWS.md
@@ -1,3 +1,6 @@
+## 2018-11-13 v1.0.1
+ * Implementation of file storing
+ * Implementation of all services and endpoints
 ## 2018-10-22 v0.0.3
  * Services skeleton was added
 ## 2018-10-22 v0.0.2